+++ src/components/ChooseDataFolder.tsx
@@ -1,10 +1,9 @@
 interface ChooseDataFolderArgs{
     setFolderPath: (path: string) => void
-    getAllProjects: (path: string) => void
-    text: string
+    text?: string
 }
 
-function ChooseDataFolder({setFolderPath, getAllProjects, text}: ChooseDataFolderArgs) {
+function ChooseDataFolder({setFolderPath, text}: ChooseDataFolderArgs) {
     const openDialog = () => {
         const result: Promise<string[] | undefined> = window.api.openFolderPathDialog();
         result.then(
@@ -12,7 +11,6 @@ function ChooseDataFolder({setFolderPath, getAllProjects, text}: ChooseDataFolde
                 if(value === undefined) return;
                 else {
                     setFolderPath(value[0].normalize());
-                    getAllProjects(value[0].normalize());
                 }
             }
         )
@@ -20,9 +18,9 @@ function ChooseDataFolder({setFolderPath, getAllProjects, text}: ChooseDataFolde
     }
 
   return (
-    <button className="btn btn-light" onClick={openDialog}>
+    <button className="btn btn-dark" onClick={openDialog}>
         {text}
-        <i className="bi bi-folder ms-2"></i>
+        <i className={"bi bi-folder " + (text !== undefined ? "ms-2" : "")}></i>
     </button>
   )
 }