+++ src/hooks/useApp.ts
@@ -5,14 +5,26 @@ interface UseAppReturn {
     error: Error | null;
     startingAPI: boolean
     isBugged: boolean
+    currentProject: string | null
+    folderPath: string | null
 
     startApi: () => Promise<void>
+    setCurrentProject: (project: string | null) => void
+    setFolder: (folderPath: string) => void
 }
 
 export function useApp(): UseAppReturn {
     const [error, setError] = useState<Error | null>(null);
     const [startingAPI, setStartingAPI] = useState<boolean>(true);
     const [isBugged, setIsBugged] = useState<boolean>(false);
+    const [currentProject, setCurrentProject] = useState<string | null>(null);
+    const [folderPath, setFolderPath] = useState<string | null>(null);
+
+    const setFolder = async (path: string): Promise<void> => {
+        apiService.updateRootFolder(path).then(
+            () => setFolderPath(path)
+        )
+    }
 
     const startApi = async (): Promise<void> => {
         const result = await apiService.startAPI();
@@ -36,5 +48,5 @@ export function useApp(): UseAppReturn {
         startApi();
     }, []);
 
-    return {startingAPI, isBugged, error, startApi}
+    return {startingAPI, isBugged, error, currentProject, folderPath, startApi, setCurrentProject, setFolder}
 }